Various fixes

+++ libopendmi/include/opendmi/backend/generic.h
@@ -13,7 +13,7 @@
 
 __BEGIN_DECLS
 
-bool dmi_generic_parse_entry_addr(dmi_context_t *context, const char *str, uint64_t *paddr);
+bool dmi_generic_parse_entry_addr(dmi_context_t *context, const char *str, size_t *paddr);
 
 #if defined(__i386__) or defined(__x86_64__)
     bool dmi_generic_find_entry_addr(
